+++ src/components/Search/SearchIconWrapper.js
@@ -1,0 +1,12 @@
+import { styled } from "@mui/material/styles";
+const SearchIconWrapper = styled("div")(({ theme }) => ({
+  padding: theme.spacing(0, 2),
+  height: "100%",
+  position: "absolute",
+  pointerEvents: "none",
+  display: "flex",
+  alignItems: "center",
+  justifyContent: "center",
+}));
+
+export default SearchIconWrapper;
