+++ src/components/Search/StyledInputBase.js
@@ -1,0 +1,12 @@
+import { styled } from "@mui/material/styles";
+import { InputBase } from "@mui/material";
+const StyledInputBase = styled(InputBase)(({ theme }) => ({
+  color: "inherit",
+  "& .MuiInputBase-input": {
+    padding: theme.spacing(1, 1, 1, 0),
+    // vertical padding + font size from searchIcon
+    paddingLeft: `calc(1em + ${theme.spacing(4)})`,
+    width: "100%",
+  },
+}));
+export default StyledInputBase;
