+++ src/components/Search/index.js
@@ -1,0 +1,3 @@
+export { default as Search } from "./Search";
+export { default as SearchIconWrapper } from "./SearchIconWrapper";
+export { default as StyledInputBase } from "./StyledInputBase";
