+++ src/pages/dashboard/Group.js
@@ -1,0 +1,107 @@
+import React, { useState } from "react";
+import {
+  Box,
+  Stack,
+  Typography,
+  IconButton,
+  Link,
+  Divider,
+} from "@mui/material";
+import { MagnifyingGlass, Plus } from "phosphor-react";
+import { useTheme } from "@mui/material/styles";
+import { SimpleBarStyle } from "../../components/Scrollbar";
+import { ChatList } from "../../data";
+import ChatElement from "../../components/ChatElement";
+import {
+  Search,
+  SearchIconWrapper,
+  StyledInputBase,
+} from "../../components/Search";
+
+const Group = () => {
+  const [openDialog, setOpenDialog] = useState(false);
+
+  const handleCloseDialog = () => {
+    setOpenDialog(false);
+  };
+  const handleOpenDialog = () => {
+    setOpenDialog(true);
+  };
+  const theme = useTheme();
+  return (
+    <>
+      <Stack direction={"row"} sx={{ width: "100%" }}>
+        <Box
+          sx={{
+            overflowY: "scroll",
+
+            height: "100vh",
+            width: 320,
+            backgroundColor: (theme) =>
+              theme.palette.mode === "light"
+                ? "#F8FAFF"
+                : theme.palette.background.paper,
+
+            boxShadow: "0px 0px 2px rgba(0, 0, 0, 0.25)",
+          }}
+        >
+          <Stack p={3} spacing={2} sx={{ maxHeight: "100vh" }}>
+            <Stack
+              alignItems={"center"}
+              justifyContent={"space-between"}
+              direction={"row"}
+            >
+              <Typography variant="h5">Groups</Typography>
+            </Stack>
+            <Stack sx={{ width: "100%" }}>
+              <Search>
+                <SearchIconWrapper>
+                  <MagnifyingGlass color="#709CE6" />
+                </SearchIconWrapper>
+                <StyledInputBase
+                  placeholder="Search…"
+                  inputProps={{ "aria-label": "search" }}
+                />
+              </Search>
+            </Stack>
+            <Stack
+              justifyContent={"space-between"}
+              alignItems={"center"}
+              direction={"row"}
+            >
+              <Typography variant="subtitle2" sx={{}} component={Link}>
+                Create New Group
+              </Typography>
+              <IconButton onClick={handleOpenDialog}>
+                <Plus style={{ color: theme.palette.primary.main }} />
+              </IconButton>
+            </Stack>
+            <Divider />
+            <Stack sx={{ flexGrow: 1, overflowY: "scroll", height: "100%" }}>
+              <SimpleBarStyle timeout={500} clickOnTrack={false}>
+                <Stack spacing={2.4}>
+                  <Typography variant="subtitle2" sx={{ color: "#676667" }}>
+                    Pinned
+                  </Typography>
+                  {/* Chat List */}
+                  {ChatList.filter((el) => el.pinned).map((el, idx) => {
+                    return <ChatElement chatData={el} />;
+                  })}
+                  <Typography variant="subtitle2" sx={{ color: "#676667" }}>
+                    All Chats
+                  </Typography>
+                  {/* Chat List */}
+                  {ChatList.filter((el) => !el.pinned).map((el, idx) => {
+                    return <ChatElement chatData={el} />;
+                  })}
+                </Stack>
+              </SimpleBarStyle>
+            </Stack>
+          </Stack>
+        </Box>
+      </Stack>
+    </>
+  );
+};
+
+export default Group;
