+++ src/routes/index.js
@@ -30,7 +30,7 @@ export default function Router() {
       children: [
         { element: <Navigate to={DEFAULT_PATH} replace />, index: true },
         { path: "app", element: <GeneralApp /> },
-
+        { path: "group", element: <GroupPage /> },
         { path: "404", element: <Page404 /> },
         { path: "*", element: <Navigate to="/404" replace /> },
       ],
@@ -42,5 +42,6 @@ export default function Router() {
 const GeneralApp = Loadable(
   lazy(() => import("../pages/dashboard/GeneralApp"))
 );
+const GroupPage = Loadable(lazy(() => import("../pages/dashboard/Group")));
 const LoginPage = Loadable(lazy(() => import("../pages/auth/Login")));
 const Page404 = Loadable(lazy(() => import("../pages/Page404")));
